-New mail verification classes -R360 project structure refactoring

+++ TestAutomationProject/src/test/java/com/cyberiansoft/test/baseutils/BaseUtils.java
@@ -34,7 +34,7 @@ public class BaseUtils {
 		return mobileplatform;
 	}
 	
-	public static void waitABit(int milliseconds) {
+	public static void waitABit(long milliseconds) {
 		try {
 			Thread.sleep(milliseconds);
 		} catch (InterruptedException e) {
